src/worker/requirements.txt
pandas==1.4.2
confluent-kafka==1.8.2